Upgrade JUnit and dependencies JUnit: 4.11 JMock: 2.6.0 cglib: 2.2.3 hamcrest: 1.3

+++ api/src/org/labkey/api/collections/CaseInsensitiveMapTest.java
@@ -127,7 +127,7 @@ public class CaseInsensitiveMapTest extends Assert
 
         for (Future<Integer> future : futures)
         {
-            assertEquals(0, future.get());
+            assertEquals(0, future.get().intValue());
         }
     }
 }
